# services/discovery/discovery_go/docs/swagger.yaml
basePath: /discovery
definitions:
  main.DiscoveryProfile:
    properties:
      bio:
        type: string
      character_class:
        type: string
      display_name:
        type: string
      gender:
        type: string
      image_urls:
        items:
          type: string
        type: array
      is_active:
        type: boolean
      profile_id:
        type: string
      realm:
        type: string
      tagline:
        type: string
      talents:
        items:
          type: string
        type: array
    type: object
  main.FeedResponse:
    properties:
      profiles:
        items:
          $ref: '#/definitions/main.DiscoveryProfile'
        type: array
    type: object
  main.MatchOut:
    properties:
      created_at:
        type: string
      id:
        type: string
      profiles:
        items:
          type: string
        type: array
    type: object
  main.SwipeCreate:
    properties:
      direction:
        type: string
      swiped_profile_id:
        type: string
      swiper_profile_id:
        type: string
    required:
    - direction
    - swiped_profile_id
    - swiper_profile_id
    type: object
  main.SwipeOut:
    properties:
      created_at:
        type: string
      direction:
        type: string
      id:
        description: Legacy field name
        type: string
      match_id:
        description: New explicit field name
        type: string
      swipe_id:
        type: string
      swiped_profile_id:
        type: string
      swiper_profile_id:
        type: string
    type: object
info:
  contact: {}
  description: Hero feed generation, swipe recording, and match management.
  title: Discovery Service API
  version: "1.0"
paths:
  /all:
    delete:
      description: Deletes all swipes and matches. Admin/test use only.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Purge all discovery data
      tags:
      - admin
  /feed/{profile_id}:
    get:
      consumes:
      - application/json
      description: Returns a curated list of active profiles the caller hasn't swiped
        on yet.
      parameters:
      - description: Caller's Profile ID
        in: path
        name: profile_id
        required: true
        type: string
      - default: 10
        description: Max profiles to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.FeedResponse'
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get discovery feed
      tags:
      - feed
  /health:
    get:
      description: Returns the health status of the discovery service.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health check
      tags:
      - health
  /matches/{id}:
    get:
      description: Returns the details of a specific match.
      parameters:
      - description: Match ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.MatchOut'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a match by ID
      tags:
      - matches
  /matches/profile/{profile_id}:
    get:
      description: Returns all matches that include the given profile ID.
      parameters:
      - description: Profile ID
        in: path
        name: profile_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.MatchOut'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List matches for a profile
      tags:
      - matches
  /swipe/:
    post:
      consumes:
      - application/json
      description: Records a left/right swipe. If mutual right-swipe is detected,
        creates a match.
      parameters:
      - description: Swipe payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/main.SwipeCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.SwipeOut'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Record a swipe
      tags:
      - swipes
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"